# wiki/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ops Agent API",
    "version": "0.1.0",
    "description": "Generated from registered Ops Agent routes. Middleware authentication, manually parsed request models and SSE media types are documented. WebSocket frames are described separately. Documentation viewer does not execute requests."
  },
  "paths": {
...
    "/api/knowledge/extractions/{conversation_id}": {
      "post": {
        "summary": "Start Knowledge Extraction",
        "operationId": "start_knowledge_extraction_api_knowledge_extractions__conversation_id__post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeGenerateDraftRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeExtractionJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      },
      "get": {
        "summary": "Knowledge Extraction Status",
        "operationId": "knowledge_extraction_status_api_knowledge_extractions__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeExtractionStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{entry_id}/versions": {
      "get": {
        "summary": "Knowledge Versions",
        "operationId": "knowledge_versions_api_knowledge__entry_id__versions_get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeVersionView"
                  },
                  "title": "Response Knowledge Versions Api Knowledge  Entry Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{entry_id}/versions/{version_id}": {
      "get": {
        "summary": "Knowledge Version Preview",
        "operationId": "knowledge_version_preview_api_knowledge__entry_id__versions__version_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeVersionPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{entry_id}/versions/{version_id}/restore": {
      "post": {
        "summary": "Restore Knowledge Version",
        "operationId": "restore_knowledge_version_api_knowledge__entry_id__versions__version_id__restore_post",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeRestoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{entry_id}/markdown": {
      "get": {
        "summary": "Knowledge Markdown",
        "operationId": "knowledge_markdown_api_knowledge__entry_id__markdown_get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMarkdownView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/from-conversation/{conversation_id}": {
      "post": {
        "summary": "Generate Knowledge Draft",
        "operationId": "generate_knowledge_draft_api_knowledge_from_conversation__conversation_id__post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeGenerateDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGenerateDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge": {
      "post": {
        "summary": "Create Knowledge Entry",
        "operationId": "create_knowledge_entry_api_knowledge_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeEntryCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      },
      "get": {
        "summary": "Search Knowledge Entries",
        "operationId": "search_knowledge_entries_api_knowledge_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "assetId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetid"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Tag"
            }
          },
          {
            "name": "sourceConversationId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sourceconversationid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/knowledge/{entry_id}": {
      "get": {
        "summary": "Get Knowledge Entry",
        "operationId": "get_knowledge_entry_api_knowledge__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      },
      "put": {
        "summary": "Update Knowledge Entry",
        "operationId": "update_knowledge_entry_api_knowledge__entry_id__put",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeEntryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      },
      "delete": {
        "summary": "Delete Knowledge Entry",
        "operationId": "delete_knowledge_entry_api_knowledge__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "knowledge"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeExtractionJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "default": "queued"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "created_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Created Ids"
          },
          "updated_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Updated Ids"
          },
          "kept_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kept Ids"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "total_batches": {
            "type": "integer",
            "title": "Total Batches",
            "default": 0
          },
          "completed_batches": {
            "type": "integer",
            "title": "Completed Batches",
            "default": 0
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "progress": {
            "type": "string",
            "title": "Progress",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeExtractionJob"
      },
      "KnowledgeExtractionStatus": {
        "properties": {
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeExtractionJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEntryView"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "KnowledgeExtractionStatus"
      },
      "KnowledgeGenerateDraftRequest": {
        "properties": {
          "maxSourceEvents": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Maxsourceevents",
            "default": 120
          },
          "modelName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelname"
          }
        },
        "type": "object",
        "title": "KnowledgeGenerateDraftRequest"
      },
...
      "KnowledgeVersionPreview": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown"
          },
          "diff": {
            "type": "string",
            "title": "Diff"
          },
          "currentUpdatedAt": {
            "type": "string",
            "title": "Currentupdatedat"
          }
        },
        "type": "object",
        "required": [
          "markdown",
          "diff",
          "currentUpdatedAt"
        ],
        "title": "KnowledgeVersionPreview"
      },
      "KnowledgeVersionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "updatedAt",
          "title"
        ],
        "title": "KnowledgeVersionView"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}